Add -o flag to write
This modifies the default dedupping behavior by always
writing content and then overwriting if it was duplicated.
This results in as much writing as there would be without
duplicates, but space is saved.

The default dedupping behavior is to read twice (once
physically and then from VFS cache), but write only once.
The default behavior is more suitable for SSDs, and the
modified behavior might be more suited for HDDs.

Either reading (always) twice, or writing (sometimes) twice
is necessary because we need to somehow calculate the hash
before we know if it is actually duplicated or not.